requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
Flask==2.3.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.1
keras==2.12.0
lxml==4.9.2
MarkupSafe==2.1.3
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.23.5
pandas==2.0.3
Pillow==10.0.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.1
Werkzeug==2.3.6
zipp==3.15.0
